Refactor AppDetailsModal and add edit chart version placeholder page.
