+++ src/components/Cluster/ClusterDetail/AppDetailsModal/EditChartVersionPane.js
@@ -1,0 +1,7 @@
+const EditChartVersionPane = () => {
+  return 'EditChartVersionPane';
+};
+
+EditChartVersionPane.propTypes = {};
+
+export default EditChartVersionPane;
